Adds a youtube remote and abstracts some code
This is still not a final design, but trying to DRY up the code.
Youtube is in and there is now a way to generate a large chunk of
controller methods as well as api tests.

I'll probably add Spotify and then call it a day on the api side to move
onto the new client repo.

I still haven't looked into CORS (which I think will be a thing for me) or
fully tested all of these commands - even though from an API/HTTP stand point
the methods are tested.