[#2] Implement get volume
Adds an endpoint to get the current volume level as an int from 0-100.
The command is tied directly to pulse and I am still unclear if that is
a bad thing or not. It will be something that potentially changes in
the future.

I also added in some more smarts for test mode. Now, the exec method
takes in an optional 2nd param which will be used to describe test data.
This is the first time I encountered this because this is the first GET
request I added to the server.

I'm also not convinced I am returning errors the way I want with HAPI.
I may try upgrading to the latest release after this and see if that helps.